Add bug fixes, oc env file.

+++ ipm/models/idct.py
@@ -359,7 +359,7 @@ class IDCT(nn.Module):
 
     def predict(self, observation):
         # [0.0137, -0.0230, -0.0459, -0.0483]
-        observation = torch.from_numpy(observation).to(self.device)
+        observation = torch.from_numpy(observation).to(self.device).float()
         observation = observation.unsqueeze(0)
         logits = self.forward(observation)
         return logits.argmax(dim=-1).cpu().numpy()[0]
